+++ src/view/table/table.view.tsx
@@ -30,6 +30,7 @@ export function tableOptions(): ViewData {
             removeCheckAll: 'boolean',
             reorderable: 'boolean',
             useCustomHeader: 'boolean',
+            defaultTypographyType: 'typographyOptions',
             headerTypographyType: {
                 default: 'title-small',
                 options: [
@@ -90,10 +91,10 @@ export function TableCreator(props: { p: TablePropsType }) {
         return {
             id: i.id,
             items: i.items,
-            checked: checkedList.includes(i.id),
+            checked: checkedList.includes(i.id.toString()),
             onCheckChange: () => {
-                if (checkedList.includes(i.id)) setCheckedList(checkedList.filter(z => z !== i.id));
-                else setCheckedList([...checkedList, i.id]);
+                if (checkedList.includes(i.id.toString())) setCheckedList(checkedList.filter(z => z !== i.id));
+                else setCheckedList([...checkedList, i.id.toString()]);
             },
             onClick: i.onClick
         }
@@ -116,18 +117,21 @@ export function TableCreator(props: { p: TablePropsType }) {
         maxHeight={props.p.stickyHeader ? "300px" : (props.p as any).withMaxHeight ? "500px" : undefined}
         pagination={(props.p as any).withPagination ? {
             endIndex: 25,
-            hasNextPage: false,
+            hasNextPage: true,
             onPageChange: () => { },
+            onPageItemCountChange: () => {},
             pageItemCount: 25,
             pageNumber: 1,
             startIndex: 1,
+            totalIndex: 100,
+            totalPageCount: 4
         } : undefined}
         isDense={props.p.isDense}
         isCheckable={(props.p as any).checkable}
         onCheckAll={(props.p as any).removeCheckAll ? undefined : v => {
             if (!v) setCheckedList([]);
             else {
-                setCheckedList(finalRows.map(i => i.id));
+                setCheckedList(finalRows.map(i => i.id.toString()));
             }
         }}
         onReorder={!(props.p as any).reorderable ? undefined : nl => {
@@ -138,6 +142,7 @@ export function TableCreator(props: { p: TablePropsType }) {
             else if (context === 'normalRows') setNormalRows(nl.map(z => z.items));
         }}
         headerTypographyType={props.p.headerTypographyType}
+        defaultTypographyType={props.p.defaultTypographyType}
     />
 
 }